+++ src/models/Isp.ts
@@ -7,6 +7,16 @@ export default class Isp extends Asn {
    */
   public readonly isp: string;
   /**
+   * The mobile country code (MCC) associated with the IP address and ISP.
+   * See https://en.wikipedia.org/wiki/Mobile_country_code.
+   */
+  public readonly mobileCountryCode?: string;
+  /**
+   * The mobile network code (MNC) associated with the IP address and ISP.
+   * See https://en.wikipedia.org/wiki/Mobile_country_code.
+   */
+  public readonly mobileNetworkCode?: string;
+  /**
    * The name of the organization associated with the IP address.
    */
   public readonly organization: string;
@@ -14,11 +24,13 @@ export default class Isp extends Asn {
   /**
    * Instantiates an "Isp" using fields from the response
    *
-   * @param response The GeoLite2 response
+   * @param response The GeoIP2-ISP response
    */
   public constructor(response: mmdb.IspResponse) {
     super(response);
     this.isp = response.isp;
+    this.mobileCountryCode = response.mobile_country_code;
+    this.mobileNetworkCode = response.mobile_network_code;
     this.organization = response.organization;
   }
 }